fix player count function

+++ functions/functions.sh
@@ -103,9 +103,17 @@ get_server_version() {
 #### Get the current connected players on a server
 
 get_connected_players() {
-    count="$(screen -S "$1" -X stuff "list\n" && sleep 0.1 && screen -S "$1" -X hardcopy /tmp/serverdump"$1".dump && tail -n2 /tmp/serverdump"$1".dump | head -n1 | grep -oP 'There are (\d+) of a max' | cut -d ' ' -f3)"
-    echo "$count"
-    rm /tmp/serverdump"$1".dump
+    running="$(is_server_running $1)"
+
+    if [[ "$running" == "false" ]]; then
+        echo "0"
+    fi
+
+    if [[ "$running" == "true" ]]; then
+        count="$(screen -S "$1" -X stuff "list\n" && sleep 0.1 && screen -S "$1" -X hardcopy /tmp/serverdump"$1".dump && tail -n2 /tmp/serverdump"$1".dump | head -n1 | grep -oP 'There are (\d+) of a max' | cut -d ' ' -f3)"
+        echo "$count"
+        rm /tmp/serverdump"$1".dump
+    fi
 }
